cemgpt-backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
faster-whisper